type adjustments and added more fine-grained control for object selections
