+++ nuxt.config.ts
@@ -1,14 +1,5 @@
-import { createResolver } from "@nuxt/kit";
-
-const {resolve} = createResolver(import.meta.url);
 export default defineNuxtConfig({
   modules: [
     '@nuxtjs/tailwindcss'
   ],
-  hooks: {
-    'prepare:types': (config) => {
-      config.tsConfig.compilerOptions.paths ||= {};
-      config.tsConfig.compilerOptions.paths['#build/soa-components'] = [resolve('./playground/.nuxt/soa-components.mjs')];
-    },
-  },
 });
