+++ soa.d.ts
@@ -1,0 +1,6 @@
+
+declare module '#build/soa-components' {
+  import type { StoryConfigProps } from "./src/types";
+  
+  export const { components = (componentNames as StoryConfigProps[]) };
+}
