+++ src/runtime/assets/css/soa-tw.css
@@ -554,20 +554,8 @@ video {
   display: none;
 }
 
-.soa-absolute {
-  position: absolute;
-}
-
 .soa-relative {
   position: relative;
-}
-
-.-soa-right-2 {
-  right: -0.5rem;
-}
-
-.soa-top-1\/2 {
-  top: 50%;
 }
 
 .soa-mx-auto {
@@ -607,16 +595,8 @@ video {
   height: 2.5rem;
 }
 
-.soa-h-8 {
-  height: 2rem;
-}
-
 .soa-h-full {
   height: 100%;
-}
-
-.soa-w-4 {
-  width: 1rem;
 }
 
 .soa-w-full {
@@ -625,14 +605,6 @@ video {
 
 .soa-grid-cols-1 {
   grid-template-columns: repeat(1, minmax(0, 1fr));
-}
-
-.soa-grid-cols-2 {
-  grid-template-columns: repeat(2, minmax(0, 1fr));
-}
-
-.soa-grid-cols-3 {
-  grid-template-columns: repeat(3, minmax(0, 1fr));
 }
 
 .soa-flex-col {
@@ -655,6 +627,10 @@ video {
   justify-content: space-between;
 }
 
+.soa-gap-10 {
+  gap: 2.5rem;
+}
+
 .soa-gap-2 {
   gap: 0.5rem;
 }
@@ -663,16 +639,12 @@ video {
   gap: 1rem;
 }
 
-.soa-gap-10 {
-  gap: 2.5rem;
+.soa-rounded-2xl {
+  border-radius: 1rem;
 }
 
 .soa-rounded-lg {
   border-radius: 0.5rem;
-}
-
-.soa-rounded-2xl {
-  border-radius: 1rem;
 }
 
 .soa-border {
@@ -715,10 +687,6 @@ video {
   background-color: var(--soa-primary);
 }
 
-.soa-bg-secondary {
-  background-color: var(--soa-secondary);
-}
-
 .soa-bg-white {
   --tw-bg-opacity: 1;
   background-color: rgb(255 255 255 / var(--tw-bg-opacity));
@@ -741,11 +709,6 @@ video {
 .soa-py-1 {
   padding-top: 0.25rem;
   padding-bottom: 0.25rem;
-}
-
-.soa-py-10 {
-  padding-top: 2.5rem;
-  padding-bottom: 2.5rem;
 }
 
 .soa-py-4 {
@@ -812,10 +775,6 @@ video {
   color: rgb(156 163 175 / var(--tw-text-opacity));
 }
 
-.hover\:soa-cursor-grab:hover {
-  cursor: grab;
-}
-
 .hover\:soa-underline:hover {
   text-decoration-line: underline;
 }
@@ -827,27 +786,11 @@ video {
   }
 }
 
-@media (min-width: 768px) {
-  .md\:soa-grid-cols-2 {
-    grid-template-columns: repeat(2, minmax(0, 1fr));
-  }
-}
-
-@media (min-width: 1024px) {
-  .lg\:soa-grid-cols-3 {
-    grid-template-columns: repeat(3, minmax(0, 1fr));
-  }
-}
-
 @media (min-width: 1280px) {
   .xl\:soa-mx-auto {
     margin-left: auto;
     margin-right: auto;
   }
-}
-
-.\[\&\:\:--webkit-slider-thumb\]\:soa-bg-primary::--webkit-slider-thumb {
-  background-color: var(--soa-primary);
 }
 
 .\[\&\:\:-webkit-slider-thumb\]\:soa-appearance-none::-webkit-slider-thumb {
